PUT FILE
The PUT method is most often used to update an existing resource.
let’s try to understand the PUT method mechanism:
In HTTP.PUT method, the resource is first identified by the URL and if it exists, then it is updated, otherwise, a new resource is created. In simply we can say that If the resource exists then update else create a new resource.